Add a pileup.destroy() method.
Hopefully this fixes #268. I believe the source of the flakiness is that
some event listeners (e.g. TwoBitDataSource.newdata) were being
triggered _after_ the pileup was removed from the DOM. This led to code
running in our React components after they were detached, which confused
React and led to the `Invariant Violation` errors.